--- src/main/java/com/evenly/took/domain/auth/client/authcode/AuthCodeRequestUrlProvider.java
@@ -1,10 +1,0 @@
-package com.evenly.took.domain.auth.client.authcode;
-
-import com.evenly.took.domain.auth.domain.OAuthType;
-
-public interface AuthCodeRequestUrlProvider {
-
-	OAuthType supportType();
-
-	String provide();
-}
