--- src/main/java/com/evenly/took/domain/auth/dto/response/JwtResponse.java
@@ -1,6 +1,0 @@
-package com.evenly.took.domain.auth.dto.response;
-
-public record JwtResponse(
-	String accessToken
-) {
-}
