+++ src/main/java/com/evenly/took/feature/auth/dto/response/JwtResponse.java
@@ -1,0 +1,6 @@
+package com.evenly.took.feature.auth.dto.response;
+
+public record JwtResponse(
+	String accessToken
+) {
+}
