+++ src/main/java/com/evenly/took/feature/user/dao/UserRepository.java
@@ -1,0 +1,13 @@
+package com.evenly.took.feature.user.dao;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.evenly.took.feature.auth.domain.OAuthIdentifier;
+import com.evenly.took.feature.user.domain.User;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+	Optional<User> findByOauthIdentifier(OAuthIdentifier oauthIdentifier);
+}
