+++ src/main/java/com/evenly/took/global/security/client/AuthCodeRequestUrlProvider.java
@@ -1,0 +1,10 @@
+package com.evenly.took.global.security.client;
+
+import com.evenly.took.feature.auth.domain.OAuthType;
+
+public interface AuthCodeRequestUrlProvider {
+
+	OAuthType supportType();
+
+	String provide();
+}
